requirements.txt
python-telegram-bot==12.3.0
spotdl==3.5.1